refactor(app): rename applyWheelZoom.js to zoom.service.ts
