+++ src/talk/talk.window.js
@@ -7,7 +7,7 @@ const { BrowserWindow, nativeTheme } = require('electron')
 const { applyExternalLinkHandler } = require('../app/externalLinkHandlers.js')
 const { applyContextMenu } = require('../app/applyContextMenu.js')
 const { applyDownloadHandler } = require('../app/downloads.ts')
-const { applyWheelZoom } = require('../app/applyWheelZoom.js')
+const { applyWheelZoom } = require('../app/zoom.service.ts')
 const { setupTray } = require('../app/app.tray.js')
 const { getBrowserWindowIcon, getTrayIcon } = require('../shared/icons.utils.js')
 const { TITLE_BAR_HEIGHT } = require('../constants.js')
